fix: user can now update profil pic when signed in with oauth

+++ back-end/src/main/java/com/braintrain/backend/controller/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     public ResponseEntity<String> updateProfilePicture(@PathVariable String userEmail, @RequestParam("file") MultipartFile file) {
         User user = userService.getUserByEmail(userEmail);
         String userProfilePictureUrl = user.getProfilePicture();
-        if (userProfilePictureUrl != null){
+        if (userProfilePictureUrl != null && userProfilePictureUrl.contains("https://storage.googleapis.com/")){
             fileService.deleteFile(userProfilePictureUrl);
         }
         return ResponseEntity.ok(userService.saveProfilePicture(userEmail, file));
